requirements.txt
asgiref==3.8.1
boto3==1.36.16
botocore==1.36.16
Brotli==1.1.0
click==8.1.8
dj-database-url==2.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
gunicorn==23.0.0
h11==0.14.0
jmespath==1.0.1
packaging==24.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.11.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
whitenoise==6.9.0
